requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-RESTful==0.3.10
Flask-Cors==4.0.0
python-dotenv==1.0.0
twilio==8.10.0
openai==1.6.1
numpy==1.26.2
pandas==2.1.3
pydub==0.25.1
SpeechRecognition==3.10.0
gunicorn==21.2.0
pytest==7.4.3
Werkzeug==2.3.7
email-validator==2.1.0
python-magic==0.4.27
requests==2.31.0
SQLAlchemy==2.0.23
Jinja2==3.1.2
cryptography==41.0.5
PyJWT==2.8.0
Pillow==10.1.0
